refactor(actor-runtime): split SpecDrivenActor module under 500 lines (ARN-179)
Directory module: mod.rs / effects.rs / tests.rs for readability ratchet.
