+++ crates/temper-actor-runtime/src/spec_actor/mod.rs
@@ -1,0 +1,361 @@
+//! SpecDrivenActor — implements the Actor trait backed by an IOA spec.
+//!
+//! Specs describe state machines (states, transitions, guards, effects).
+//! The routing is external — reaction rules wire emit effects to target actors.
+//!
+//! # Architecture
+//!
+//! - Spec → TransitionTable (via temper-jit)
+//! - Reaction rules → routing map (emit name → target actor type)
+//! - handle(): evaluate table → apply effects → route emits via ctx.tell()
+//!
+//! # Message protocol
+//!
+//! Actors communicate via `SpecMessage { action, params }`:
+//! - `action`: the action/emit name (e.g., "PrepareContext")
+//! - `params`: JSON-encoded params (empty for actions with no params)
+
+use std::collections::{BTreeMap, HashMap};
+
+use temper_jit::table::{EvalContext, TransitionTable};
+use temper_runtime::reaction::ReactionRule;
+use temper_spec::automaton::Automaton;
+
+use crate::actor::{Actor, ActorContext, ActorError, ActorHandle, Message};
+
+mod effects;
+
+// ─── SpecMessage ─────────────────────────────────────────────────────────────
+
+/// Generic message for spec-driven actor communication.
+/// The action name matches the IOA spec action/emit name.
+#[derive(Clone, PartialEq, prost::Message)]
+pub struct SpecMessage {
+    /// The action/emit name (e.g., "PrepareContext", "ToolCallBatchRequested").
+    #[prost(string, tag = "1")]
+    pub action: String,
+    /// JSON-encoded params (empty bytes for parameterless actions).
+    #[prost(bytes, tag = "2")]
+    pub params: Vec<u8>,
+}
+
+impl SpecMessage {
+    pub fn new(action: impl Into<String>) -> Self {
+        Self {
+            action: action.into(),
+            params: Vec::new(),
+        }
+    }
+
+    pub fn with_params(action: impl Into<String>, params: serde_json::Value) -> Self {
+        Self {
+            action: action.into(),
+            params: serde_json::to_vec(&params).unwrap_or_default(),
+        }
+    }
+}
+
+// ─── Actor state ─────────────────────────────────────────────────────────────
+
+/// Serializable state for spec-driven actors.
+#[derive(Debug, Clone, serde::Serialize, serde::Deserialize, Default)]
+pub struct SpecActorState {
+    pub status: String,
+    #[serde(default)]
+    pub counters: BTreeMap<String, usize>,
+    #[serde(default)]
+    pub booleans: BTreeMap<String, bool>,
+    #[serde(default)]
+    pub lists: BTreeMap<String, Vec<String>>,
+    /// Arbitrary extra data — used to thread params through the reaction chain.
+    /// SpecDrivenActor stores the last incoming message params here so integrations
+    /// can read them from the trigger message.
+    #[serde(default)]
+    pub fields: serde_json::Value,
+}
+
+impl SpecActorState {
+    fn to_eval_context(&self) -> EvalContext {
+        let mut ctx = EvalContext::default();
+        for (k, v) in &self.counters {
+            ctx.counters.insert(k.clone(), *v);
+        }
+        for (k, v) in &self.booleans {
+            ctx.booleans.insert(k.clone(), *v);
+        }
+        for (k, v) in &self.lists {
+            ctx.lists.insert(k.clone(), v.clone());
+        }
+        ctx
+    }
+}
+
+// ─── Routing map builder ─────────────────────────────────────────────────────
+
+/// Build per-actor routing maps from reaction rules.
+///
+/// Returns `HashMap<actor_type, HashMap<emit_name, (target_actor_type, target_action)>>`.
+pub fn build_routing_maps(
+    rules: &[ReactionRule],
+) -> HashMap<String, HashMap<String, (String, String)>> {
+    let mut maps: HashMap<String, HashMap<String, (String, String)>> = HashMap::new();
+
+    for rule in rules {
+        if let Some(emit_name) = &rule.when.action {
+            maps.entry(rule.when.entity_type.clone())
+                .or_default()
+                .insert(
+                    emit_name.clone(),
+                    (rule.then.entity_type.clone(), rule.then.action.clone()),
+                );
+        }
+    }
+
+    maps
+}
+
+/// Build a single actor's routing map from a reaction registry.
+pub fn build_actor_routing(
+    actor_type: &str,
+    rules: &[ReactionRule],
+) -> HashMap<String, (String, String)> {
+    rules
+        .iter()
+        .filter(|r| r.when.entity_type == actor_type)
+        .filter_map(|r| {
+            r.when.action.as_ref().map(|emit| {
+                (
+                    emit.clone(),
+                    (r.then.entity_type.clone(), r.then.action.clone()),
+                )
+            })
+        })
+        .collect()
+}
+
+// ─── SpecDrivenActor ─────────────────────────────────────────────────────────
+
+/// An Actor implementation driven by an IOA spec + reaction routing.
+///
+/// - State machine transitions validated by the TransitionTable
+/// - Emit effects routed to sibling actors via ctx.tell()
+/// - Trigger effects sent to integration actors via ctx.tell()
+pub struct SpecDrivenActor {
+    /// Actor type name (e.g., "Agent", "ContextManager").
+    pub(crate) name: String,
+    /// TransitionTable compiled from the IOA spec.
+    pub(crate) table: TransitionTable,
+    /// Initial state (from spec's initial state + variable declarations).
+    pub(crate) init_state: SpecActorState,
+    /// Routing map: emit/trigger name → (target actor type, target action).
+    pub(crate) routing: HashMap<String, (String, String)>,
+    /// Leaked static refs for subscriptions() return.
+    pub(crate) subscriptions_static: Vec<&'static str>,
+}
+
+impl SpecDrivenActor {
+    /// Create from an IOA TOML source + routing map.
+    ///
+    /// Fails closed (ADR-0168) if the compiled transition table uses effects
+    /// this backend cannot execute (`schedule`, `schedule_at`, `spawn`).
+    pub fn from_ioa(
+        ioa_source: &str,
+        routing: HashMap<String, (String, String)>,
+    ) -> Result<Self, String> {
+        let automaton = temper_spec::parse_automaton(ioa_source)
+            .map_err(|e| format!("failed to parse spec: {e}"))?;
+        Self::from_automaton(&automaton, ioa_source, routing)
+    }
+
+    /// Create from a pre-parsed Automaton + routing map.
+    ///
+    /// Returns `Err` when the table contains schedule/spawn effects that the
+    /// PG actor-runtime cannot honor (ADR-0168 / ARN-179).
+    pub fn from_automaton(
+        automaton: &Automaton,
+        ioa_source: &str,
+        routing: HashMap<String, (String, String)>,
+    ) -> Result<Self, String> {
+        let name = automaton.automaton.name.clone();
+        let table = TransitionTable::from_ioa_source(ioa_source);
+        effects::reject_unsupported_effects(&table)?;
+
+        // Build initial state from spec variables.
+        let mut init_state = SpecActorState {
+            status: automaton.automaton.initial.clone(),
+            ..Default::default()
+        };
+        for var in &automaton.state {
+            match var.var_type.as_str() {
+                "counter" => {
+                    let v: usize = var.initial.parse().unwrap_or(0);
+                    init_state.counters.insert(var.name.clone(), v);
+                }
+                "bool" => {
+                    let v: bool = var.initial.parse().unwrap_or(false);
+                    init_state.booleans.insert(var.name.clone(), v);
+                }
+                "list" | "set" => {
+                    init_state.lists.insert(var.name.clone(), Vec::new());
+                }
+                _ => {}
+            }
+        }
+
+        // Input actions are the message types this actor accepts.
+        // NOTE: Box::leak is intentional — actors are singletons, never dropped.
+        let subscriptions_static: Vec<&'static str> = automaton
+            .actions
+            .iter()
+            .filter(|a| a.kind == "input")
+            .map(|a| &*Box::leak(a.name.clone().into_boxed_str()))
+            .collect();
+
+        Ok(Self {
+            name,
+            table,
+            init_state,
+            routing,
+            subscriptions_static,
+        })
+    }
+
+    /// Which message types this actor accepts.
+    pub fn subscription_strings(&self) -> &[&'static str] {
+        &self.subscriptions_static
+    }
+
+    /// The routing map (emit name → target actor type).
+    pub fn routing(&self) -> &HashMap<String, (String, String)> {
+        &self.routing
+    }
+}
+
+#[async_trait::async_trait]
+impl Actor for SpecDrivenActor {
+    fn actor_type(&self) -> &str {
+        &self.name
+    }
+
+    fn initial_state(&self) -> Vec<u8> {
+        serde_json::to_vec(&self.init_state).unwrap_or_default()
+    }
+
+    async fn handle(
+        &self,
+        ctx: &ActorContext,
+        state: &mut Vec<u8>,
+        message: &Message,
+    ) -> Result<(), ActorError> {
+        // 1. Deserialize state.
+        let mut actor_state: SpecActorState = if state.is_empty() {
+            self.init_state.clone()
+        } else {
+            serde_json::from_slice(state)
+                .map_err(|e| ActorError::HandlerFailed(format!("state deser: {e}")))?
+        };
+
+        // 2. Resolve action name + params.
+        // If the message carries a SpecMessage, extract the action from its payload.
+        // This handles both direct SpecMessage sends and raw action-name messages.
+        let spec_msg = if message.message_type.ends_with("SpecMessage") {
+            message.decode::<SpecMessage>().ok()
+        } else {
+            None
+        };
+        let action = spec_msg
+            .as_ref()
+            .filter(|m| !m.action.is_empty())
+            .map(|m| m.action.clone())
+            .unwrap_or_else(|| message.message_type.clone());
+
+        // Store incoming params in state.fields so integrations can read them.
+        // Merge non-empty params into fields to preserve context from prior steps
+        // (e.g. child Process keeps parent_pid while later messages add user_prompt/response).
+        // For a new user turn, clear transient scratchpad fields from prior turns.
+        if self.name == "Process"
+            && matches!(action.as_str(), "StartProcess" | "SendInput")
+            && let Some(obj) = actor_state.fields.as_object_mut()
+        {
+            for key in [
+                "tool_calls",
+                "tool_results",
+                "child_result",
+                "response",
+                "error",
+            ] {
+                obj.remove(key);
+            }
+        }
+
+        if let Some(fields) = spec_msg
+            .as_ref()
+            .filter(|m| !m.params.is_empty())
+            .and_then(|m| serde_json::from_slice::<serde_json::Value>(&m.params).ok())
+            .filter(|p| !p.as_object().is_some_and(|o| o.is_empty()))
+        {
+            match (actor_state.fields.as_object_mut(), fields.as_object()) {
+                (Some(existing), Some(new_fields)) => {
+                    for (k, v) in new_fields {
+                        existing.insert(k.clone(), v.clone());
+                    }
+                }
+                _ => actor_state.fields = fields,
+            }
+        }
+
+        let eval_ctx = actor_state.to_eval_context();
+
+        // 2. Evaluate transition table.
+        let result = self
+            .table
+            .evaluate_ctx(&actor_state.status, &eval_ctx, &action);
+
+        match result {
+            Some(r) if r.success => {
+                let from_status = actor_state.status.clone();
+
+                // 3. Apply effects — may include SetState.
+                for effect in &r.effects {
+                    self.apply_effect_inner(&mut actor_state, effect, ctx).await;
+                }
+
+                // 4. Apply state transition fallback (if no SetState effect fired).
+                if actor_state.status == from_status && !r.new_state.is_empty() {
+                    actor_state.status = r.new_state.clone();
+                }
+
+                tracing::info!(
+                    actor = %self.name,
+                    action = %action,
+                    new_state = %actor_state.status,
+                    "transition"
+                );
+            }
+            Some(_) => {
+                tracing::warn!(
+                    actor = %self.name,
+                    action = %action,
+                    status = %actor_state.status,
+                    "action not valid from current state"
+                );
+            }
+            None => {
+                tracing::warn!(
+                    actor = %self.name,
+                    action = %action,
+                    "unknown action"
+                );
+            }
+        }
+
+        // 5. Serialize state back.
+        *state = serde_json::to_vec(&actor_state)
+            .map_err(|e| ActorError::HandlerFailed(format!("state ser: {e}")))?;
+
+        Ok(())
+    }
+}
+
+#[cfg(test)]
+mod tests;
